refactor: introduce dependency injection
Follow the instructions in section "Adding Dependency Injection" of
https://levelup.gitconnected.com/create-a-quick-net-core-cli-with-system-commandline-logging-and-dependency-injection-e369b4491f78
